hybrid config added, tests passing
