+++ src/Serilog.Sinks.MSSqlServer/Configuration/Implementations/SetProperty.cs
