+++ src/Serilog.Sinks.MSSqlServer/Configuration/Implementations/System.Configuration/ApplySystemConfiguration.cs
@@ -1,143 +1,23 @@
-﻿// Copyright 2014 Serilog Contributors
-// 
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// 
-//     http://www.apache.org/licenses/LICENSE-2.0
-// 
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using System;
+﻿using System;
 using System.Collections.ObjectModel;
-using System.Data;
 using Serilog.Configuration;
-using Serilog.Events;
-using Serilog.Sinks.MSSqlServer;
 using System.Configuration;
 using Serilog.Debugging;
 
-namespace Serilog
+namespace Serilog.Sinks.MSSqlServer
 {
     /// <summary>
-    /// Adds the WriteTo.MSSqlServer() extension method to <see cref="LoggerConfiguration"/>.
+    /// Configures the sink's connection string and ColumnOtions object.
     /// </summary>
-    public static class LoggerConfigurationMSSqlServerExtensions
+    internal static class ApplySystemConfiguration
     {
-        /// <summary>
-        /// The configuration section name for app.config or web.config configuration files.
-        /// </summary>
-        public static string AppConfigSectionName = "MSSqlServerSettingsSection";
-
-        /// <summary>
-        /// Adds a sink that writes log events to a table in a MSSqlServer database.
-        /// Create a database and execute the table creation script found here
-        /// https://gist.github.com/mivano/10429656
-        /// or use the autoCreateSqlTable option.
-        /// </summary>
-        /// <param name="loggerConfiguration">The logger configuration.</param>
-        /// <param name="connectionString">The connection string to the database where to store the events.</param>
-        /// <param name="tableName">Name of the table to store the events in.</param>
-        /// <param name="schemaName">Name of the schema for the table to store the data in. The default is 'dbo'.</param>
-        /// <param name="restrictedToMinimumLevel">The minimum log event level required in order to write an event to the sink.</param>
-        /// <param name="batchPostingLimit">The maximum number of events to post in a single batch.</param>
-        /// <param name="period">The time to wait between checking for event batches.</param>
-        /// <param name="formatProvider">Supplies culture-specific formatting information, or null.</param>
-        /// <param name="autoCreateSqlTable">Create log table with the provided name on destination sql server.</param>
-        /// <param name="columnOptions"></param>
-        /// <returns>Logger configuration, allowing configuration to continue.</returns>
-        /// <exception cref="ArgumentNullException">A required parameter is null.</exception>
-        public static LoggerConfiguration MSSqlServer(
-            this LoggerSinkConfiguration loggerConfiguration,
-            string connectionString,
-            string tableName,
-            LogEventLevel restrictedToMinimumLevel = LevelAlias.Minimum,
-            int batchPostingLimit = MSSqlServerSink.DefaultBatchPostingLimit,
-            TimeSpan? period = null,
-            IFormatProvider formatProvider = null,
-            bool autoCreateSqlTable = false,
-            ColumnOptions columnOptions = null,
-            string schemaName = "dbo"
-            )
-        {
-            if (loggerConfiguration == null) throw new ArgumentNullException("loggerConfiguration");
-
-            var defaultedPeriod = period ?? MSSqlServerSink.DefaultPeriod;
-
-            if (ConfigurationManager.GetSection(AppConfigSectionName) is MSSqlServerConfigurationSection serviceConfigSection)
-                columnOptions = ConfigureColumnOptions(serviceConfigSection, columnOptions);
-
-            connectionString = GetConnectionString(connectionString);
-
-            return loggerConfiguration.Sink(
-                new MSSqlServerSink(
-                    connectionString,
-                    tableName,
-                    batchPostingLimit,
-                    defaultedPeriod,
-                    formatProvider,
-                    autoCreateSqlTable,
-                    columnOptions,
-                    schemaName
-                    ),
-                restrictedToMinimumLevel);
-        }
-
-        /// <summary>
-        /// Adds a sink that writes log events to a table in a MSSqlServer database.
-        /// Create a database and execute the table creation script found here
-        /// https://gist.github.com/mivano/10429656
-        /// or use the autoCreateSqlTable option.
-        /// </summary>
-        /// <param name="loggerAuditSinkConfiguration">The logger configuration.</param>
-        /// <param name="connectionString">The connection string to the database where to store the events.</param>
-        /// <param name="tableName">Name of the table to store the events in.</param>
-        /// <param name="schemaName">Name of the schema for the table to store the data in. The default is 'dbo'.</param>
-        /// <param name="restrictedToMinimumLevel">The minimum log event level required in order to write an event to the sink.</param>
-        /// <param name="formatProvider">Supplies culture-specific formatting information, or null.</param>
-        /// <param name="autoCreateSqlTable">Create log table with the provided name on destination sql server.</param>
-        /// <param name="columnOptions"></param>
-        /// <returns>Logger configuration, allowing configuration to continue.</returns>
-        /// <exception cref="ArgumentNullException">A required parameter is null.</exception>
-        public static LoggerConfiguration MSSqlServer(this LoggerAuditSinkConfiguration loggerAuditSinkConfiguration,
-                                                      string connectionString,
-                                                      string tableName,
-                                                      LogEventLevel restrictedToMinimumLevel = LevelAlias.Minimum,
-                                                      IFormatProvider formatProvider = null,
-                                                      bool autoCreateSqlTable = false,
-                                                      ColumnOptions columnOptions = null,
-                                                      string schemaName = "dbo")
-        {
-            if (loggerAuditSinkConfiguration == null) throw new ArgumentNullException("loggerAuditSinkConfiguration");
-
-            if (ConfigurationManager.GetSection(AppConfigSectionName) is MSSqlServerConfigurationSection serviceConfigSection)
-                columnOptions = ConfigureColumnOptions(serviceConfigSection, columnOptions);
-
-            connectionString = GetConnectionString(connectionString);
-
-            return loggerAuditSinkConfiguration.Sink(
-                new MSSqlServerAuditSink(
-                    connectionString,
-                    tableName,
-                    formatProvider,
-                    autoCreateSqlTable,
-                    columnOptions,
-                    schemaName
-                    ),
-                restrictedToMinimumLevel);
-        }
-
         /// <summary>
         /// Examine if supplied connection string is a reference to an item in the "ConnectionStrings" section of web.config
         /// If it is, return the ConnectionStrings item, if not, return string as supplied.
         /// </summary>
         /// <param name="nameOrConnectionString">The name of the ConnectionStrings key or raw connection string.</param>
         /// <remarks>Pulled from review of Entity Framework 6 methodology for doing the same</remarks>
-        private static string GetConnectionString(string nameOrConnectionString)
+        internal static string GetConnectionString(string nameOrConnectionString)
         {
 
             // If there is an `=`, we assume this is a raw connection string not a named value
@@ -161,7 +41,7 @@ namespace Serilog
         /// <summary>
         /// Populate ColumnOptions properties and collections from app config
         /// </summary>
-        private static ColumnOptions ConfigureColumnOptions(MSSqlServerConfigurationSection config, ColumnOptions columnOptions)
+        internal static ColumnOptions ConfigureColumnOptions(MSSqlServerConfigurationSection config, ColumnOptions columnOptions)
         {
             var opts = columnOptions ?? new ColumnOptions();
 
@@ -203,7 +83,7 @@ namespace Serilog
                 {
                     foreach (ColumnConfig c in config.Columns)
                     {
-                        if(!string.IsNullOrWhiteSpace(c.ColumnName))
+                        if (!string.IsNullOrWhiteSpace(c.ColumnName))
                         {
                             if (opts.AdditionalColumns == null)
                                 opts.AdditionalColumns = new Collection<SqlColumn>();
