+++ src/Serilog.Sinks.MSSqlServer/Serilog.Sinks.MSSqlServer.csproj
@@ -4,7 +4,7 @@
     <Description>A Serilog sink that writes events to Microsoft SQL Server</Description>
     <VersionPrefix>5.1.3</VersionPrefix>
     <Authors>Michiel van Oudheusden;Serilog Contributors</Authors>
-    <TargetFrameworks>netstandard2.0;net45</TargetFrameworks>
+    <TargetFrameworks>netstandard2.0;net452;netcoreapp2.0;net461</TargetFrameworks>
     <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <AssemblyName>Serilog.Sinks.MSSqlServer</AssemblyName>
@@ -25,39 +25,54 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <!-- Whenever these are updated, also update versions in packages.config for net452 -->
     <PackageReference Include="Serilog" Version="2.5.0" />
     <PackageReference Include="Serilog.Sinks.PeriodicBatching" Version="2.1.1" />
   </ItemGroup>
 
   <ItemGroup>
     <!-- Ensure MSBuild ignores all build-target-specific files by default -->
-    <Compile Remove="Configuration\NetFramework.ConfigurationManager\**\*.*" />
-    <Compile Remove="Configuration\Microsoft.Extensions.Configuration\**\*.*" />
+    <Compile Remove="Configuration\Extensions\Hybrid\**\*.*" />
+    <Compile Remove="Configuration\Extensions\Microsoft.Extensions.Configuration\**\*.*" />
+    <Compile Remove="Configuration\Extensions\System.Configuration\**\*.*" />
+    <Compile Remove="Configuration\Implementations\Microsoft.Extensions.Configuration\**\*.*" />
+    <Compile Remove="Configuration\Implementations\System.Configuration\**\*.*" />
     <!-- Show in VStudio, but MSBuild ignores these (indicates files are not code, non-published-content, etc.) -->
-    <None Include="Configuration\NetFramework.ConfigurationManager\**\*.*" />
-    <None Include="Configuration\Microsoft.Extensions.Configuration\**\*.*" />
+    <None Include="Configuration\Extensions\Hybrid\**\*.*" />
+    <None Include="Configuration\Extensions\Microsoft.Extensions.Configuration\**\*.*" />
+    <None Include="Configuration\Extensions\System.Configuration\**\*.*" />
+    <None Include="Configuration\Implementations\Microsoft.Extensions.Configuration\**\*.*" />
+    <None Include="Configuration\Implementations\System.Configuration\**\*.*" />
     <!-- ItemGroups below with TFM conditions will re-include the compile targets -->
   </ItemGroup>
  
-  <ItemGroup Condition=" '$(TargetFramework)' == 'netstandard2.0'">
+  <ItemGroup Condition=" '$(TargetFramework)' == 'netstandard2.0' ">
     <PackageReference Include="System.Data.SqlClient" Version="4.4.0" />
     <PackageReference Include="Microsoft.Extensions.Configuration" Version="2.0.0" />
     <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="2.0.0" />
-    <Compile Include="Configuration\Microsoft.Extensions.Configuration\**\*.cs" />
+    <Compile Include="Configuration\Extensions\Microsoft.Extensions.Configuration\**\*.cs" />
+    <Compile Include="Configuration\Implementations\Microsoft.Extensions.Configuration\**\*.cs" />
   </ItemGroup>
-  
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net45' ">
+
+  <ItemGroup Condition=" '$(TargetFramework)' == 'netcoreapp2.0' Or '$(TargetFramework)' == 'net461' ">
+    <PackageReference Include="System.Configuration.ConfigurationManager" Version="4.5.0" />
+    <PackageReference Include="System.Data.SqlClient" Version="4.4.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration" Version="2.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="2.0.0" />
+    <Compile Include="Configuration\Extensions\Hybrid\**\*.cs" />
+    <Compile Include="Configuration\Implementations\Microsoft.Extensions.Configuration\**\*.cs" />
+    <Compile Include="Configuration\Implementations\System.Configuration\**\*.cs" />
+  </ItemGroup>
+
+  <ItemGroup Condition=" '$(TargetFramework)' == 'net452' ">
     <Reference Include="System.Data" />
     <Reference Include="System.Xml" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Configuration" />
     <Reference Include="System" />
     <Reference Include="Microsoft.CSharp" />
-    <Compile Include="Configuration\NetFramework.ConfigurationManager\**\*.cs" />
+    <Compile Include="Configuration\Extensions\System.Configuration\**\*.cs" />
+    <Compile Include="Configuration\Implementations\System.Configuration\**\*.cs" />
   </ItemGroup>
-  
-  <ItemGroup>
-    <None Include="Configuration\SetProperty.cs" />
-  </ItemGroup>
-  
+
 </Project>
