+++ src/Serilog.Sinks.MSSqlServer/packages.config
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Serilog" version="2.5.0" targetFramework="net45" />
-  <package id="Serilog.Sinks.PeriodicBatching" version="2.0.0-beta-700" targetFramework="net45" />
+  <package id="Serilog" version="2.5.0" targetFramework="net452" />
+  <package id="Serilog.Sinks.PeriodicBatching" version="2.1.1" targetFramework="net452" />
 </packages>
